requirements.txt
albumentations==2.0.8
bing_image_downloader==1.1.2
captum==0.8.0
diffusers==0.34.0
feedparser==6.0.11
fpdf==1.7.2
huggingface_hub==0.34.4
matplotlib==3.10.5
numpy==1.26.4
openai-clip==1.0.1
openai==1.101.0
opencv-python-headless==4.10.0.84
pandas==2.3.2
Pillow==11.3.0
pydeck==0.9.1
python-dotenv==1.1.1
pytesseract==0.3.13
requests==2.32.5
roboflow==1.2.1
scikit-learn==1.7.1
seaborn==0.13.2
streamlit==1.46.0
tensorboard==2.15.0
torch==2.7.1
torchmetrics==1.7.3
torchvision==0.22.1
tqdm==4.67.1
transformers==4.55.4
wandb==0.21.1
wikipedia==1.4.0